backend/requirements.txt
fastapi
uvicorn
pydantic
gspread
google-auth
pandas
python-dotenv
